+++ test/test_mazegraph.cpp
@@ -21,7 +21,7 @@ TEST(fourwaygraph, unit)
     EXPECT_EQ(0, id_from = mg.nodeIdByAgentState({ { 0, 0 }, kNoDirection, 0 }));
     EXPECT_EQ(16, id_to = mg.nodeIdByAgentState({ { 0, 2 }, kNoDirection, 0 }));
     EXPECT_EQ(1, mg.distance(id_from, id_to));
-    EXPECT_EQ(1, mg.distance(id_from + 1, id_to));
+    EXPECT_EQ(1, mg.distance(uint16_t(id_from + 1), id_to));
 
     std::vector<uint16_t> v;
     mg.neighbors(0, v);
